+++ pred_src/tawhiri_axes.h
@@ -1,7 +1,7 @@
 #ifndef TAWHIRI_AXES_H
 #define TAWHIRI_AXES_H
 
-typedef double dataset_array_t[65][47][3][361][720];
+typedef float dataset_array_t[65][47][3][361][720];
 extern const int shape[];
 extern const int dimensions;
 extern const int axis_0_hour[];
